+++ BriansUsbQuizBoxApi/QuizBoxTypeEnum.cs
@@ -1,0 +1,17 @@
+﻿using System;
+using System.ComponentModel;
+
+namespace BriansUsbQuizBoxApi
+{
+    /// <summary>
+    /// Quiz box enumerations
+    /// </summary>
+    public enum QuizBoxTypeEnum
+    {
+        [Description("Brian's Quiz Box")]
+        BriansQuizBox = 1,
+
+        [Description("Kirkman Basic Quizbox")]
+        KirkmanQuizBox = 2
+    }
+}